# Conflicts:
#	logs/UpdateStatut.log

+++ src/main/java/fr/eni/encheres/service/ArticleService.java
@@ -47,7 +47,8 @@ public interface ArticleService {
     @Transactional
     int updateECtoVD();
 
-    long finalPointInProgress(long id_article);
+
+    int finalPointInProgress(int id_article);
 }
